# api-service/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "api-service",
		"description": "guggr api definition",
		"contact": {
			"name": "Guggr"
		},
		"license": {
			"name": ""
		},
		"version": "0.1.0"
	},
	"paths": {
...
		"/api/v1/users": {
			"get": {
				"tags": ["users"],
				"summary": "list endpoint for users",
				"operationId": "list_user",
				"responses": {
					"200": {
						"description": "List users",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/DisplayUser"
									}
								}
							}
						}
					},
					"500": {
						"description": "Storage error"
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			},
			"post": {
				"tags": ["users"],
				"summary": "create endpoint for users",
				"operationId": "create_user",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUser"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Created user",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DisplayUser"
								}
							}
						}
					},
					"400": {
						"description": "Validation error"
					},
					"500": {
						"description": "Storage error"
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			}
		},
		"/api/v1/users/{id}": {
			"get": {
				"tags": ["users"],
				"summary": "get endpoint for users",
				"operationId": "get_user",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "User id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "User",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DisplayUser"
								}
							}
						}
					},
					"500": {
						"description": "Storage error"
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			},
			"delete": {
				"tags": ["users"],
				"summary": "delete endpoint for users",
				"operationId": "delete_user",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "User id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Deleted"
					},
					"500": {
						"description": "Storage error"
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			},
			"patch": {
				"tags": ["users"],
				"summary": "update endpoint for users",
				"operationId": "update_user",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "User id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUser"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Patched user",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DisplayUser"
								}
							}
						}
					},
					"400": {
						"description": "Validation error"
					},
					"500": {
						"description": "Storage error"
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			}
		}
	},
	"components": {
		"schemas": {
...
			"CreateUser": {
				"type": "object",
				"description": "sent to the database for creating a new user",
				"required": ["name", "email", "password"],
				"properties": {
					"email": {
						"type": "string",
						"format": "email"
					},
					"name": {
						"type": "string",
						"minLength": 1
					},
					"password": {
						"type": "string",
						"format": "password",
						"minLength": 8
					}
				}
			},
...
			"DisplayUser": {
				"type": "object",
				"description": "returned from the database when displaying a user",
				"required": ["id", "name", "email"],
				"properties": {
					"email": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				}
			},
...
			"UpdateUser": {
				"type": "object",
				"description": "sent to the database when updating a user",
				"properties": {
					"email": {
						"type": ["string", "null"],
						"format": "email"
					},
					"name": {
						"type": ["string", "null"],
						"minLength": 1
					},
					"password": {
						"type": ["string", "null"],
						"format": "password",
						"minLength": 8
					}
				}
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	}
}